+++ sdk/core/azure-core/test/ut/transport_adapter_base.hpp
@@ -19,6 +19,36 @@
 
 namespace Azure { namespace Core { namespace Test {
 
+  namespace _detail {
+    constexpr static const char AzureSdkHttpbinServerSchema[] = "https://";
+    constexpr static const char AzureSdkHttpbinServer[] = "azuresdkforcpp.azurewebsites.net";
+  } // namespace _detail
+
+  struct AzureSdkHttpbinServer
+  {
+    inline static std::string Get()
+    {
+      return std::string(_detail::AzureSdkHttpbinServerSchema)
+          + std::string(_detail::AzureSdkHttpbinServer) + "/get";
+    }
+    inline static std::string Put()
+    {
+      return std::string(_detail::AzureSdkHttpbinServerSchema)
+          + std::string(_detail::AzureSdkHttpbinServer) + "/put";
+    }
+    inline static std::string Delete()
+    {
+      return std::string(_detail::AzureSdkHttpbinServerSchema)
+          + std::string(_detail::AzureSdkHttpbinServer) + "/delete";
+    }
+    inline static std::string Patch()
+    {
+      return std::string(_detail::AzureSdkHttpbinServerSchema)
+          + std::string(_detail::AzureSdkHttpbinServer) + "/patch";
+    }
+    inline static std::string Host() { return std::string(_detail::AzureSdkHttpbinServer); }
+  };
+
   struct TransportAdaptersTestParameter
   {
     std::string Suffix;